pluginlink: HookEvent fails hard on failure, add HookEventOptionally
This commit changes the default HookEvent() implementation to fail
hard if no event of that name was available.

This is mostly for lazy plugin authors (like me) that want to hook
events and just fail when they don't exist, because the plugin would
not work otherwise. It is also a very sane default to have.

For people who are still looking for optional events, a
HookEventOptionally() macro is provided that works like HookEvent() did
before.

Best used in OnPluginsLoaded or similar - fail early and hard, like you
read about.

This commit also patches the builtin plugins to use the new call
semantics. (all two of them!)

Internally, this adds a new entry to pluginLink (at the end), so this
should not break existing plugins except those that depended on explicit
checking.